pipelines/requirements.txt
google-cloud-storage==1.43.0
gcsfs==0.9.1
pandas==1.3.3
pandas-gbq==0.17.0
google-cloud-bigquery==2.26.0
python-dateutil==2.8.2
